requirements.txt
argostranslate==1.9.1
docx==0.2.4
eparse==0.7.3
extra_streamlit_components==0.1.71
fastapi==0.111.0
langchain==0.2.1
langchain_community==0.2.1
langchain_core==0.2.3
langchain_google_genai==1.0.5
langfuse==2.33.1
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
protobuf==5.27.0
psycopg2==2.9.9
pydantic==2.7.3
PyJWT==2.8.0
PyJWT==2.8.0
python-dotenv==1.0.1
python_bcrypt==0.3.2
python_docx==1.1.0
SQLAlchemy==2.0.27
streamlit==1.31.0
streamlit_authenticator==0.3.1
streamlit_extras==0.4.3
streamlit_option_menu==0.3.12
streamlit_sortables==0.2.0
